+++ DryAgentSystem/DryAgentSystem/Models/RateRequest.cs
@@ -85,7 +85,7 @@ namespace DryAgentSystem.Models
 
         [Display(Name = "Gross Weight")]
         [Range(0.00001, int.MaxValue, ErrorMessage = "The field {0} must be greater than 0")]
-        public decimal GrossWt { get; set; }
+        public double GrossWt { get; set; }
 
         [Display(Name = "Gross Weight Unit")]
         public string GrossWtUnit { get; set; }
@@ -132,13 +132,13 @@ namespace DryAgentSystem.Models
 
         [Display(Name = "Rate (USD)")]
         [Range(0.00001,int.MaxValue, ErrorMessage = "The field {0} must be greater than 0")]
-        public decimal Rate { get; set; }
+        public double Rate { get; set; }
 
         [Display(Name = "Rate Countered")]
-        public decimal RateCountered { get; set; }
+        public double RateCountered { get; set; }
 
         [Display(Name = "Grand Total Sales")]
-        public decimal GTotalSalesCal { get; set; }
+        public double GTotalSalesCal { get; set; }
 
         [Display(Name = "Rate ID")]
         public string RateID { get; set; }
